requirements.txt
django
django-crispy-forms
crispy-bootstrap5
pillow